+++ src/main/java/ru/vsu/entity/mappers/ParamsMapper.java
@@ -9,9 +9,9 @@ import java.sql.SQLException;
 public class ParamsMapper implements RowMapper<ParamsEntity> {
     @Override
     public ParamsEntity mapRow(ResultSet resultSet, int i) throws SQLException {
-        return new ParamsEntity(resultSet.getBigDecimal("id").toBigInteger(),
-                resultSet.getBigDecimal("attr_id").toBigInteger(),
-                resultSet.getBigDecimal("object_id").toBigInteger(),
+        return new ParamsEntity(resultSet.getLong("id"),
+                resultSet.getLong("attr_id"),
+                resultSet.getLong("object_id"),
                 resultSet.getString("value"));
     }
